pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arable"
version = "0.2.0"
description = "Agentic Runtime And Business Logic Engine"
authors = [{name = "Ian Armstrong", email = "ian@wowitsian.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "textual>=0.44.0",
    "pydantic>=2.0.0", 
    "anthropic>=0.8.0",
    "requests>=2.28.0",
    "gspread>=5.0.0",
    "oauth2client>=4.1.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
    "python-dotenv>=0.19.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pytest-cov>=4.0.0"
]

[project.scripts]
arable = "arable.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["arable*"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
